[AWIT-92-zapisywanie-danych-uzytkownika-w-bazie-danych] Moved response/request records to body_models package
