+++ backend/src/main/java/pl/edu/agh/awiteks_backend/api/auth/AuthController.java
@@ -7,6 +7,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import pl.edu.agh.awiteks_backend.api.auth.body_models.AuthData;
+import pl.edu.agh.awiteks_backend.api.auth.body_models.UserLoginRequestBody;
+import pl.edu.agh.awiteks_backend.api.auth.body_models.UserRegisterRequestBody;
 import pl.edu.agh.awiteks_backend.services.AuthService;
 
 @RestController
