+++ backend/src/main/java/pl/edu/agh/awiteks_backend/api/auth/body_models/BadRegisterData.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.awiteks_backend.api.auth;
+package pl.edu.agh.awiteks_backend.api.auth.body_models;
 
 import io.swagger.v3.oas.annotations.media.Schema;
