+++ backend/src/main/java/pl/edu/agh/awiteks_backend/api/auth/body_models/UserLoginRequestBody.java
@@ -1,9 +1,8 @@
-package pl.edu.agh.awiteks_backend.api.auth;
+package pl.edu.agh.awiteks_backend.api.auth.body_models;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-public record UserRegisterRequestBody(
-        @Schema(required = true) String username,
+public record UserLoginRequestBody(
         @Schema(required = true) String email,
         @Schema(required = true) String password
 ) {
