+++ backend/src/main/java/pl/edu/agh/awiteks_backend/api/auth/body_models/UserRegisterRequestBody.java
@@ -1,8 +1,9 @@
-package pl.edu.agh.awiteks_backend.api.auth;
+package pl.edu.agh.awiteks_backend.api.auth.body_models;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-public record UserLoginRequestBody(
+public record UserRegisterRequestBody(
+        @Schema(required = true) String username,
         @Schema(required = true) String email,
         @Schema(required = true) String password
 ) {
