+++ backend/src/main/java/pl/edu/agh/awiteks_backend/api/plants/body_models/AddPlantRequestBody.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.awiteks_backend.api.plants;
+package pl.edu.agh.awiteks_backend.api.plants.body_models;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import pl.edu.agh.awiteks_backend.models.Insolation;
