+++ backend/src/main/java/pl/edu/agh/awiteks_backend/api/plants/body_models/PlantsStats.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.awiteks_backend.api.plants;
+package pl.edu.agh.awiteks_backend.api.plants.body_models;
 
 import io.swagger.v3.oas.annotations.media.Schema;
