+++ backend/src/main/java/pl/edu/agh/awiteks_backend/api/species/body_models/AddSpeciesRequestBody.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.awiteks_backend.api.species;
+package pl.edu.agh.awiteks_backend.api.species.body_models;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import pl.edu.agh.awiteks_backend.models.Fertilization;
