+++ backend/src/main/java/pl/edu/agh/awiteks_backend/api/users/body_models/UserInfo.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.awiteks_backend.api.users;
+package pl.edu.agh.awiteks_backend.api.users.body_models;
 
 import io.swagger.v3.oas.annotations.media.Schema;
