+++ backend/src/main/java/pl/edu/agh/awiteks_backend/services/AuthService.java
@@ -8,10 +8,10 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import pl.edu.agh.awiteks_backend.api.auth.AuthData;
-import pl.edu.agh.awiteks_backend.api.auth.BadRegisterData;
-import pl.edu.agh.awiteks_backend.api.auth.UserLoginRequestBody;
-import pl.edu.agh.awiteks_backend.api.auth.UserRegisterRequestBody;
+import pl.edu.agh.awiteks_backend.api.auth.body_models.AuthData;
+import pl.edu.agh.awiteks_backend.api.auth.body_models.BadRegisterData;
+import pl.edu.agh.awiteks_backend.api.auth.body_models.UserLoginRequestBody;
+import pl.edu.agh.awiteks_backend.api.auth.body_models.UserRegisterRequestBody;
 import pl.edu.agh.awiteks_backend.models.User;
 import pl.edu.agh.awiteks_backend.repositories.UserRepository;
 import pl.edu.agh.awiteks_backend.security.jwt.JwtAccessToken;
